libarchive: Cast constants to int64_t instead of using LL suffix
The LL suffix is not portable.  Use an explicit cast instead.

+++ Utilities/cmlibarchive/libarchive/archive_read_support_format_iso9660.c
@@ -1406,7 +1406,7 @@ zisofs_read_data(struct archive_read *a,
 
 		/* Allocate block pointers buffer. */
 		ceil = (zisofs->pz_uncompressed_size +
-			(1LL << zisofs->pz_log2_bs) - 1)
+			(((int64_t)1) << zisofs->pz_log2_bs) - 1)
 			>> zisofs->pz_log2_bs;
 		xsize = (ceil + 1) * 4;
 		if (zisofs->block_pointers_alloc < xsize) {
